move b2b entity into b2b namespace
